Frankly, I'm not sure why the two casts I've adopted here are useful, as I wasn't getting clang warnings on them, but whatever. Also, did not adopt the bug in the exponent printing change (our repo already has that taken care of anyway.)

+++ printf.c
@@ -383,12 +383,12 @@ static size_t _ftoa(out_fct_type out, char* buffer, size_t idx, size_t maxlen, d
   int whole = (int)value;
   double tmp = (value - whole) * pow10[prec];
   unsigned long frac = (unsigned long)tmp;
-  diff = tmp - frac;
+  diff = tmp - (double)frac;
 
   if (diff > 0.5) {
     ++frac;
     // handle rollover, e.g. case 0.99 with prec 1 is 1.0
-    if (frac >= pow10[prec]) {
+    if ((double)frac >= pow10[prec]) {
       frac = 0;
       ++whole;
     }
